EqipmentDropdown added to container folder
