+++ src/components/EquipmentDropdown/EquipmentDropdown.css
