+++ src/components/EquipmentDropdown/index.js
@@ -1,0 +1,16 @@
+import { connect } from 'react-redux';
+import { modifyEquipment } from '../../actions';
+import EquipmentDropdown from './component';
+import './EquipmentDropdown.css';
+
+
+const mapStateToProps = state => ({
+  currentView: state.currentView,
+  totalWeight: state.totalWeight,
+  characterStats: state.characterStats,
+  gear: {
+    equipment: state.gear.equipment,
+  },
+});
+
+export default connect(mapStateToProps, { modifyEquipment })(EquipmentDropdown);
